CI-4705
we already have a colleague who want to ims module of this sdk (outside of the events context).

So I think we would be better of with com.adobe as a group
and aio-lib-java as a repo and parent project name on github.
I updated the tickets and comments above accordingly